Add new members to an existing cluster one by one.
Before sending add member command to an existing cluster we should check
that there is nobody already in process of adding itself to the cluster.

Basically this is just workaround for the following problem:
https://github.com/zalando/stups-etcd-cluster/issues/1

New members were added successfully but etcd failed to start due to
version incompatibility and we lost quorum.