Add Route53 Records to the Stack for etcd.
Currently, DNS records maintained by the etcd-cluster will not be removed when the whole CloudFormation stack is
removed. This means manual cleanup is necessary when an etcd-cluster is removed.
By adding the records to the CloudFormation stack for etcd, we ensure they are also cleaned up nicely when
a stack is removed.